refactor desktópe
