+++ src/catalog/catalog.module.ts
@@ -1,6 +1,6 @@
+import { DataModule } from "@data/data.module";
 import { Schemas } from "@data/data.schemas";
 import { Module } from "@nestjs/common";
-import { DataModule } from "@data/data.module";
 
 import { CatalogController } from "./catalog.controller";
 import { ComposeController } from "./compose.controller";
